requirements.txt
Django>=5.0
Pillow>=10.0
openpyxl>=3.1.0
